requirements.txt
Django==1.11
djangorestframework==3.9.0
Markdown==3.0.1
pandas==0.23.4